# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Ayoitson API
  version: 1.0.0
  description: |
    The Ayoitson REST API. Phase 4 redesign — uniform POST/PUT/DELETE
    semantics across every resource, structured error envelope on every
    response, X-API-Key auth on every endpoint except `GET /api/health` and
    `POST /api/auth/setup`. See `docs/SDLC.md` and the vault `[[API Redesign]]`
    note for the design memo.
  contact:
    name: SALM LLC / Ayoitson
servers:
  - url: http://localhost:8000
    description: Local development
security:
  - apiKey: []
...
paths:
...
  /api/settings/ffmpeg:
    get:
      tags: [settings]
      summary: Read FFmpeg settings
      responses:
        '200':
          description: FFmpeg settings
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
        '401': { $ref: '#/components/responses/Unauthorized' }
    put:
      tags: [settings]
      summary: Update FFmpeg settings
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, additionalProperties: true }
      responses:
        '200':
          description: Updated FFmpeg settings
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/settings/ffmpeg/known-paths:
    get:
      tags: [settings]
      summary: List detected well-known FFmpeg paths
      responses:
        '200':
          description: Absolute FFmpeg paths detected on this host
          content:
            application/json:
              schema:
                type: object
                required: [paths]
                properties:
                  paths:
                    type: array
                    items: { type: string }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/settings/ffmpeg/reset:
    post:
      tags: [settings]
      summary: Reset FFmpeg settings to defaults
      responses:
        '200':
          description: Reset FFmpeg settings
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/settings/plex:
    get:
      tags: [settings]
      summary: Read Plex settings
      responses:
        '200': { description: Plex settings, content: { application/json: { schema: { type: object, additionalProperties: true } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
    put:
      tags: [settings]
      summary: Update Plex settings
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, additionalProperties: true }
      responses:
        '200': { description: Updated Plex settings, content: { application/json: { schema: { type: object, additionalProperties: true } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/settings/plex/reset:
    post:
      tags: [settings]
      summary: Reset Plex settings to defaults
      responses:
        '200': { description: Reset Plex settings, content: { application/json: { schema: { type: object, additionalProperties: true } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/settings/xmltv:
    get:
      tags: [settings]
      summary: Read XMLTV settings
      description: |
        Note — the XMLTV file path is hard-coded server-side at `${DATABASE}/xmltv.xml`.
        The `file` field is intentionally omitted from request and response shapes.
        This closes F10-xmltv-readfile (operator-controllable arbitrary-file-read).
      responses:
        '200':
          description: XMLTV settings (no `file` field)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/XmltvSettings' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    put:
      tags: [settings]
      summary: Update XMLTV settings
      description: Bodies containing a `file` field are rejected with 400.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/XmltvSettings' }
      responses:
        '200':
          description: Updated XMLTV settings
          content:
            application/json:
              schema: { $ref: '#/components/schemas/XmltvSettings' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/settings/xmltv/reset:
    post:
      tags: [settings]
      summary: Reset XMLTV settings to defaults
      responses:
        '200':
          description: Reset XMLTV settings
          content:
            application/json:
              schema: { $ref: '#/components/schemas/XmltvSettings' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/settings/hdhr:
    get:
      tags: [settings]
      summary: Read HDHR settings
      responses:
        '200': { description: HDHR settings, content: { application/json: { schema: { type: object, additionalProperties: true } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
    put:
      tags: [settings]
      summary: Update HDHR settings
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, additionalProperties: true }
      responses:
        '200': { description: Updated HDHR settings, content: { application/json: { schema: { type: object, additionalProperties: true } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/settings/hdhr/reset:
    post:
      tags: [settings]
      summary: Reset HDHR settings to defaults
      responses:
        '200': { description: Reset HDHR settings, content: { application/json: { schema: { type: object, additionalProperties: true } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
components:
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        Issued by the first-run setup flow (`scripts/first-run.js` or
        `POST /api/auth/setup`). Format: `ayo_<base64url>`. Verified server-
        side via argon2id against the `api_keys` SQLite table.
...
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            - VALIDATION_ERROR
            - UNAUTHORIZED
            - FORBIDDEN
            - NOT_FOUND
            - CONFLICT
            - PAYLOAD_TOO_LARGE
            - RATE_LIMITED
            - INTERNAL
            - UPSTREAM_ERROR
            - SERVICE_UNAVAILABLE
        message: { type: string }
        details:
          type: object
          additionalProperties: true
...
    XmltvSettings:
      type: object
      properties:
        _id: { type: string }
        cache: { type: integer }
        refresh: { type: integer }
        enableImageCache: { type: boolean }
      additionalProperties: false
...
  responses:
    Unauthorized:
      description: Missing or invalid X-API-Key
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    ValidationError:
      description: Request body or parameter is invalid
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...